Refactor: Centralize gateway logic into ChipGateway class
